chore: refactor injectDB to inject-db
