+++ hono/src/services/items.ts
@@ -5,8 +5,8 @@ import { createInsertSchema, createSelectSchema } from 'drizzle-zod'
 import { Hono } from 'hono'
 import { z } from 'zod'
 import { items, itemsToOutfits, itemTypeEnum, outfits } from '../schema'
-import type { Variables } from '../utils/injectDB'
-import injectDB from '../utils/injectDB'
+import type { Variables } from '../utils/inject-db'
+import injectDB from '../utils/inject-db'
 
 const app = new Hono<{ Variables: Variables }>()
